port fastlookup to 2.77test4
The basic idea is use a placeholder in daemon->servers, whenever it need to
match --server/--address, instead of iterate through the linked list, dnsmasq
will look it up on daemon->htree_special_domains, then copy the match result
to the placeholder.

Ipsets match is simpler, the match result can be used directly.

The code is re-organized to use upstream functions/blocks when possible, and
most newly added code blocks are separated from current upstream source.
Hopefully, it will be easier for others to review.